Day 8 Project - Caesar Cipher
The program encodes or decodes your words that entered by shifting the letters by the amount of number you want , so that you create 'secret' understandable messages between the people you are messaging with. You need to enter encode to mess the word up, and decode to make it readable again.